requirements.txt
antiorm==1.2.1
click==8.1.3
colorama==0.4.6
db==0.1.1
db-sqlite3==0.0.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
numpy==1.24.3
pandas==2.0.2
protobuf==3.20.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
typing-extensions==4.5.0
tzdata==2023.3
Werkzeug==2.2.2
zipp==3.12.1
